pyproject.toml
[project]
name = "ask-gpt"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai (>=1.77.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "rich (>=14.0.0,<15.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
ask-gpt = "ask_gpt.cli:main"